Move ImputationReasonCode enum definition
