+++ src/pudl/analysis/timeseries_cleaning.py
@@ -34,7 +34,6 @@ import uuid
 import warnings
 from collections.abc import Callable, Sequence
 from dataclasses import dataclass
-from enum import Enum, unique
 from typing import Any, Optional
 
 import matplotlib.pyplot as plt
@@ -46,6 +45,7 @@ from dagster import AssetIn, AssetOut, asset, multi_asset
 from pandera.typing import DataFrame, Index, Series
 
 from pudl.logging_helpers import get_logger
+from pudl.metadata.enums import ImputationReasonCodes
 
 logger = get_logger(__file__)
 
@@ -69,27 +69,6 @@ mapped to their standard-time UTC offset.
 
 
 # --- Data structures --- #
-
-
-@unique
-class ImputationReasonCodes(Enum):
-    """Defines all reasons a value might be flagged for imputation."""
-
-    ANOMALOUS_REGION = "Indicates that value is surrounded by flagged values."
-    NEGATIVE_OR_ZERO = "Indicates value is negative or zero value."
-    IDENTICAL_RUN = (
-        "Indicates value is among last values in an identical run of values."
-    )
-    GLOBAL_OUTLIER = (
-        "Indicates value is greater or less than n times the global median."
-    )
-    GLOBAL_OUTLIER_NEIGHBOR = "Indicates value is a neighbors global outliers."
-    LOCAL_OUTLIER_HIGH = "Indicates value is a local outlier on the high end."
-    LOCAL_OUTLIER_LOW = "Indicates value is a local outlier on the low end."
-    DOUBLE_DELTA = "Indicates value is very different from neighbors on either side."
-    SINGLE_DELTA = (
-        "Indicates value is significantly different from nearest unflagged value."
-    )
 
 
 class UTCTimeseriesDataFrame(pa.DataFrameModel):
